一括で追加ボタンを追加 #42
